Move ClosedCfd out of sqlite module
